+++ components/inappnotification/impl/src/main/java/com/flipperdevices/inappnotification/impl/composable/ComposableInAppNotification.kt
@@ -96,7 +96,8 @@ private fun ComposableInAppNotificationCard(
                 )
 
                 is InAppNotification.Error -> ComposableInAppNotificationError(
-                    notification, onClickAction
+                    notification,
+                    onClickAction
                 )
             }
         }