+++ components/notification/impl/src/main/java/com/flipperdevices/notification/api/FlipperAppNotificationApiImpl.kt
@@ -100,6 +100,7 @@ class FlipperAppNotificationApiImpl @Inject constructor(
         }
     }
 
+    @Suppress("LongMethod")
     private suspend fun setSubscribeToUpdate(
         isSubscribe: Boolean,
         onRetry: () -> Unit,