+++ components/notification/impl/src/main/java/com/flipperdevices/notification/model/ChannelBlockedException.kt
@@ -1,4 +1,3 @@
 package com.flipperdevices.notification.model
 
-class ChannelBlockedException : Throwable() {
-}
+class ChannelBlockedException : Throwable()