+++ components/notification/impl/src/main/java/com/flipperdevices/notification/utils/NotificationPermissionHelper.kt
@@ -5,13 +5,11 @@ import android.app.NotificationManager
 import android.content.Context
 import android.content.pm.PackageManager
 import android.os.Build
-import androidx.core.app.NotificationManagerCompat
 import androidx.core.content.ContextCompat
 import com.flipperdevices.core.permission.api.PermissionRequestHandler
 import com.flipperdevices.notification.model.NotificationPermissionState
-import javax.inject.Inject
 import kotlinx.coroutines.CompletableDeferred
-
+import javax.inject.Inject
 
 class NotificationPermissionHelper @Inject constructor(
     private val context: Context,
@@ -52,5 +50,4 @@ class NotificationPermissionHelper @Inject constructor(
         }
         return NotificationPermissionState.GRANTED
     }
-
 }