+++ components/notification/impl/src/main/java/com/flipperdevices/notification/viewmodel/NotificationDialogViewModel.kt
@@ -12,8 +12,8 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.runBlocking
-import javax.inject.Provider
 import tangle.viewmodel.VMInject
+import javax.inject.Provider
 
 class NotificationDialogViewModel @VMInject constructor(
     settingsProvider: Provider<DataStore<Settings>>,